voktyabr/requirement.txt
# Core
Django==1.10
Pillow==3.0.1

# Batteries
django-mptt==0.8.6
django-mptt-admin==0.4.0
django-ckeditor==5.1.0
django-redactor==0.2
django-wysiwyg-redactor
django-redactoreditor==1.2.10
django-tinymce==2.3.0
django-suit-redactor==0.0.4


# REST

# Deploy

# dbmailer

# Other